Move auth repositories from user module to the authentication module
