+++ modules/authentication/server-scala/src/main/scala/repositories/FacebookAuthRepository.scala
@@ -1,9 +1,9 @@
-package repositories.auth
+package repositories
 
 import com.byteslounge.slickrepo.repository.Repository
 import javax.inject.Inject
-import model.oauth.facebook.FacebookAuthTable.FacebookAuthTable
-import model.oauth.facebook.FacebookAuth
+import model.facebook.FacebookAuth
+import model.facebook.FacebookAuthTable.FacebookAuthTable
 import slick.ast.BaseTypedType
 import slick.jdbc.JdbcProfile
