+++ modules/authentication/server-scala/src/main/scala/repositories/GithubAuthRepository.scala
@@ -1,9 +1,9 @@
-package repositories.auth
+package repositories
 
 import com.byteslounge.slickrepo.repository.Repository
 import javax.inject.Inject
-import model.oauth.github.GithubAuth
-import model.oauth.github.GithubAuthTable.GithubAuthTable
+import model.github.GithubAuth
+import model.github.GithubAuthTable.GithubAuthTable
 import slick.ast.BaseTypedType
 import slick.jdbc.JdbcProfile
